requirements.txt
atlassian-python-api>=3.41.0
click>=8.1.0
markdownify>=0.11.0
python-dotenv>=1.0.0
pyyaml>=6.0
